fix: handle webworker load
