+++ apps/cowswap-frontend/src/modules/analytics/index.ts
@@ -1,2 +1,2 @@
 export * from './events'
-export * from './useAnalyticsReporter'
+export * from './useAnalyticsReporterCowSwap'
