+++ libs/ui/src/analytics/useAnalyticsReporter.ts
@@ -1,12 +1,15 @@
 import { useEffect } from 'react'
+import { isMobile } from '@cowprotocol/common-utils'
 
 import { AnalyticsContext, PixelEvent, CowAnalytics, PixelAnalytics, WebVitalsAnalytics } from '@cowprotocol/analytics'
 import { usePrevious } from '@cowprotocol/common-hooks'
 import { useWalletDetails, useWalletInfo } from '@cowprotocol/wallet'
 
 import { useLocation } from 'react-router-dom'
+import { serviceWorkerLoad } from './events'
 
 let initiatedPixel = false
+let initiated = false
 
 interface UseAnalyticsReporterProps {
   cowAnalytics: CowAnalytics
@@ -26,9 +29,29 @@ export function useAnalyticsReporter(props: UseAnalyticsReporterProps) {
   const { walletName } = useWalletDetails()
   const prevAccount = usePrevious(account)
 
-  // Report Web Vitals
   useEffect(() => {
+    if (initiated) {
+      return
+    }
+    initiated = true
+
+    // Report Web Vitals
     webVitalsAnalytics?.reportWebVitals()
+
+    // Set browser type
+    if (typeof window !== 'undefined') {
+      cowAnalytics.setContext(
+        AnalyticsContext.customBrowserType,
+        !isMobile ? 'desktop' : 'web3' in window || 'ethereum' in window ? 'mobileWeb3' : 'mobileRegular'
+      )
+    }
+
+    // Report service worker status
+    if (typeof window !== 'undefined') {
+      const installed = Boolean(window.navigator.serviceWorker?.controller)
+      const hit = Boolean((window as any).__isDocumentCached)
+      serviceWorkerLoad(cowAnalytics, installed, hit)
+    }
   }, [webVitalsAnalytics])
 
   // Set analytics context: chainId
